requirements.txt
asgiref==3.7.2
Django==4.2.1
django-filter==23.2
djangorestframework==3.14.0
gunicorn==21.2.0
packaging==23.2
Pillow==10.0.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
python-slugify==8.0.1
pytz==2023.3
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.6.3
